requirements.txt
django
psycopg2-binary
djangorestframework
django-filter
python-decouple
Pillow
drf-yasg